+++ src/FourteenSegDisplay.h
@@ -36,6 +36,19 @@ public:
     
     // Methods
 
+    /*
+    Note: 
+    The digits/startPos parameter in the functions refers to the position/starting bit, for model2 only
+    0x80  b10000000 D8 LHS
+    0x40- b01000000 D7
+    0x20- b00100000 D6  
+    0x10- b00010000 D5
+    0x08- b00001000 D4
+    0x04- b00000100 D3
+    0x02- b00000010 D2
+    0x01- b00000001 D1 RHS
+    */
+    
     // Send ASCII value to 14 segment display, passed ASCII value byte, 
     // and for Model2 only digit position 0x80-0x40-0x20-0x10-0x08-0x04-0x02-0x01 D8-D1
     void displayASCII(uint8_t ascii, uint8_t digits = 0x00);